server/requirements.txt
flask~=1.1
grpcio-tools~=1.30
Jinja2~=2.11
pytest~=5.4
MarkupSafe==2.0.1
flask-wtf>=1.0.1 
